cleanup: move ShouldRetryVolumeGeneration from internal/rbd to internal/util
